requirements.txt
django-environ==0.12.0
Django==4.2.19
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-allauth==65.4.1
django-cors-headers==4.7.0
flake8==7.1.2
isort==6.0.1
black==25.1.0
numpy==1.26.4
requests==2.32.3
cryptography==44.0.2
langchain==0.3.19
langchain-chroma==0.2.2
langchain-community==0.3.18
langchain-core==0.3.39
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
pymupdf==1.25.3
daphne==4.1.2
channels==4.2.0
channels_redis==4.2.1
google-auth==2.38.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
celery==5.4.0
crewai==0.108.0
django-celery-beat==2.7.0